# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Files Exchanger API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/files/{fileId}/link_tokens": {
            "get": {
                "tags": [
                    "FileRelationships",
                    "LinkTokens"
                ],
                "summary": "Get all link tokens of file.",
                "description": "Retreive all link tokens by file id.",
                "operationId": "getFileLinkTokens",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/header-accept-type"
                    },
                    {
                        "$ref": "#/components/parameters/header-authorization-token"
                    },
                    {
                        "$ref": "#/components/parameters/file-id-path-parameter"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful get operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FileLinkTokensRelatedResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication failed. Bearer token mismatch."
                    },
                    "403": {
                        "description": "Forbidden. Probably you are trying to access link tokens of not your file."
                    },
                    "404": {
                        "description": "File with specified id not found."
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "FileRelationships",
                    "LinkTokens"
                ],
                "summary": "Store new file.",
                "description": "Put file to storage and save it's information to database.",
                "operationId": "storeNewFile",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/header-accept-type"
                    },
                    {
                        "$ref": "#/components/parameters/header-authorization-token"
                    },
                    {
                        "$ref": "#/components/parameters/file-id-path-parameter"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/create-link-token-request-body"
                },
                "responses": {
                    "200": {
                        "description": "Link token was successfully created.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/LinkTokenResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication failed. Bearer token mismatch."
                    },
                    "403": {
                        "description": "Forbidden. You are not allowed to create link tokens for this file."
                    },
                    "422": {
                        "description": "Unprocessable Entity. Given data is invalid.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": "{\"link_type\": [ \"The selected link type is invalid.\" ]}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            }
        },
        "/api/files/{fileId}/relationships/user": {
            "get": {
                "tags": [
                    "FileRelationships",
                    "Users"
                ],
                "summary": "Retreive link and user id, which file belongs to.",
                "operationId": "getFileUserRelationship",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/header-accept-type"
                    },
                    {
                        "$ref": "#/components/parameters/header-authorization-token"
                    },
                    {
                        "$ref": "#/components/parameters/file-id-path-parameter"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful get operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FileUserRelationshipResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication failed. Bearer token mismatch."
                    },
                    "404": {
                        "description": "File with specified id not found."
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            }
        },
        "/api/files/{fileId}/relationships/link_tokens": {
            "get": {
                "tags": [
                    "FileRelationships",
                    "LinkTokens"
                ],
                "summary": "Get all link token ids related to provided file.",
                "description": "Retreive all link token ids by file id.",
                "operationId": "getFileLinkTokensRelationship",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/header-accept-type"
                    },
                    {
                        "$ref": "#/components/parameters/header-authorization-token"
                    },
                    {
                        "$ref": "#/components/parameters/file-id-path-parameter"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful get operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FileLinkTokensRelationshipResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication failed. Bearer token mismatch."
                    },
                    "403": {
                        "description": "Forbidden. Probably you are trying to access link tokens of not your file."
                    },
                    "404": {
                        "description": "File with specified id not found."
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            }
        },
        "/api/files/": {
            "post": {
                "tags": [
                    "Files"
                ],
                "summary": "Store new file.",
                "description": "Put file to storage and save it's information to database.",
                "operationId": "storeNewFile",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/header-accept-type"
                    },
                    {
                        "$ref": "#/components/parameters/header-authorization-token"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/store-file-request-body"
                },
                "responses": {
                    "200": {
                        "description": "File was successfully inserted.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/FileIdentifierResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication failed. Bearer token mismatch."
                    },
                    "403": {
                        "description": "Forbidden. You are not allowed to create files."
                    },
                    "422": {
                        "description": "Unprocessable Entity. Given data is invalid.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": "{\"file\": [ \"The file field is required.\" ]}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "FileIdentifierResource": {
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "files"
                    },
                    "id": {
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
            "FileLinkTokensRelatedResource": {
                "properties": {
                    "links": {
                        "properties": {
                            "self": {
                                "type": "string",
                                "example": "http://localhost:8000/api/files/1/link_tokens"
                            }
                        },
                        "type": "object"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LinkTokenResource"
                        }
                    }
                },
                "type": "object"
            },
            "FileLinkTokensRelationshipResource": {
                "properties": {
                    "links": {
                        "properties": {
                            "self": {
                                "type": "string",
                                "example": "http://localhost:8000/api/files/1/relationships/link_tokens"
                            },
                            "related": {
                                "type": "string",
                                "example": "http://localhost:8000/api/files/1/link_tokens"
                            }
                        },
                        "type": "object"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LinkTokenIdentifierResource"
                        }
                    }
                },
                "type": "object"
            },
...
            "FileUserRelationshipResource": {
                "properties": {
                    "links": {
                        "properties": {
                            "self": {
                                "type": "string",
                                "example": "http://localhost:8000/api/files/1/relationships/user"
                            },
                            "related": {
                                "type": "string",
                                "example": "http://localhost:8000/api/users/1"
                            }
                        },
                        "type": "object"
                    },
                    "data": {
                        "$ref": "#/components/schemas/UserIdentifierResource"
                    }
                },
                "type": "object"
            },
            "LinkTokenIdentifierResource": {
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "link_tokens"
                    },
                    "id": {
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
            "LinkTokenResource": {
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "link_tokens"
                    },
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "attributes": {
                        "properties": {
                            "token": {
                                "type": "string",
                                "example": "9194e805-0408-4184-8ad1-22744d1ffe17"
                            },
                            "type": {
                                "type": "string",
                                "enum": [
                                    "disposable",
                                    "unlimited"
                                ]
                            },
                            "created_at": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "updated_at": {
                                "type": "string",
                                "format": "date-time"
                            }
                        },
                        "type": "object"
                    },
                    "links": {
                        "properties": {
                            "self": {
                                "type": "string",
                                "example": "http://localhost:8000/api/link_tokens/1"
                            },
                            "resource": {
                                "type": "string",
                                "example": "http://localhost:8000/api/guest/files/9194e805-0408-4184-8ad1-22744d1ffe17"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "UserIdentifierResource": {
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "users"
                    },
                    "id": {
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
            "header-accept-type": {
                "name": "Accept",
                "in": "header",
                "description": "Accept type",
                "required": true,
                "schema": {
                    "type": "string",
                    "example": "application/json"
                }
            },
            "header-authorization-token": {
                "name": "Authorization",
                "in": "header",
                "description": "Api authorization user token",
                "required": true,
                "schema": {
                    "type": "string",
                    "example": "Bearer 9194773b-3f24-42bb-93ca-654557dd303c"
                }
            },
            "file-id-path-parameter": {
                "name": "fileId",
                "in": "path",
                "description": "File id",
                "required": true,
                "schema": {
                    "type": "integer",
                    "example": "1"
                }
            },
...
        },
        "requestBodies": {
            "store-file-request-body": {
                "description": "File and it's information.",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "public_name",
                                "description",
                                "file"
                            ],
                            "properties": {
                                "public_name": {
                                    "type": "string",
                                    "example": "Some name"
                                },
                                "description": {
                                    "type": "string",
                                    "example": "Sensitive Description"
                                },
                                "date_to_delete": {
                                    "type": "string",
                                    "format": "date"
                                },
                                "file": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "create-link-token-request-body": {
                "description": "Link token information.",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "link_type"
                            ],
                            "properties": {
                                "link_type": {
                                    "type": "string",
                                    "enum": [
                                        "disposable",
                                        "unlimited"
                                    ]
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "bearer_auth": {
                "type": "http",
                "name": "bearer_auth",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}